[RF] Unify `chi2FitTo` and `createChi2` overloads and documentation
Before, the docs were scattered around different overloads, making it
easy to miss the documentation for the commands that were only included
in one of the docstring (like for `DataError`).
