+++ roofit/roofitcore/inc/RooAbsReal.h
@@ -180,15 +180,17 @@ public:
   bool getForceNumInt() const { return _forceNumInt ; }
 
   // Chi^2 fits to histograms
-  virtual RooFit::OwningPtr<RooFitResult> chi2FitTo(RooDataHist& data, const RooCmdArg& arg1={},  const RooCmdArg& arg2={},
-                              const RooCmdArg& arg3={},  const RooCmdArg& arg4={}, const RooCmdArg& arg5={},
-                              const RooCmdArg& arg6={},  const RooCmdArg& arg7={}, const RooCmdArg& arg8={}) ;
-  virtual RooFit::OwningPtr<RooFitResult> chi2FitTo(RooDataHist& data, const RooLinkedList& cmdList) ;
+  template <typename... CmdArgs_t>
+  RooFit::OwningPtr<RooFitResult> chi2FitTo(RooDataHist& data, CmdArgs_t const&... cmdArgs)
+  {
+    return RooFit::makeOwningPtr(chi2FitToImpl(data, *RooFit::Detail::createCmdList(&cmdArgs...)));
+  }
 
-  virtual RooFit::OwningPtr<RooAbsReal> createChi2(RooDataHist& data, const RooLinkedList& cmdList) ;
-  virtual RooFit::OwningPtr<RooAbsReal> createChi2(RooDataHist& data, const RooCmdArg& arg1={},  const RooCmdArg& arg2={},
-             const RooCmdArg& arg3={},  const RooCmdArg& arg4={}, const RooCmdArg& arg5={},
-             const RooCmdArg& arg6={},  const RooCmdArg& arg7={}, const RooCmdArg& arg8={}) ;
+  template <typename... CmdArgs_t>
+  RooFit::OwningPtr<RooAbsReal> createChi2(RooDataHist& data, CmdArgs_t const&... cmdArgs)
+  {
+    return RooFit::makeOwningPtr(createChi2Impl(data, *RooFit::Detail::createCmdList(&cmdArgs...)));
+  }
 
   // Chi^2 fits to X-Y datasets
   virtual RooFit::OwningPtr<RooFitResult> chi2FitTo(RooDataSet& xydata, const RooCmdArg& arg1={},  const RooCmdArg& arg2={},
@@ -547,6 +549,9 @@ protected:
 
 
 private:
+
+  std::unique_ptr<RooAbsReal> createChi2Impl(RooDataHist& data, const RooLinkedList& cmdList);
+  std::unique_ptr<RooFitResult> chi2FitToImpl(RooDataHist& data, const RooLinkedList& cmdList);
 
   /// Debug version of getVal(), which is slow and does error checking.
   double _DEBUG_getVal(const RooArgSet* normalisationSet) const;
