+++ tutorials/roofit/roofit/rf602_chi2fit.C
@@ -60,7 +60,7 @@ void rf602_chi2fit()
    // by the number of events in the dataset to obtain the fit function
    // If model is an extended pdf, the expected number events is used
    // instead of the observed number of events.
-   model.chi2FitTo(*dh, {PrintLevel(-1)});
+   model.chi2FitTo(*dh, PrintLevel(-1));
 
    // NB: It is also possible to fit a RooAbsReal function to a RooDataHist
    // using chi2FitTo().
